moved the translation to another page + code cleaning
