+++ app/i18n/strings_en.json
@@ -199,20 +199,20 @@
     "APPEARANCE":"Appearance",
     "UI_CONFIG_LANGUAGE":"Language",
     "UI_CONFIG_SELECT_LANGUAGE":"Select Language",
-	"UI_CONFIG_SELECT_LANGUAGE_DOC":"Select the Language of your system",
+	  "UI_CONFIG_SELECT_LANGUAGE_DOC":"Select the Language of your system",
     "NEW_LANGUAGE_SET":"New language set",
     "NEW_BACKGROUND_APPLIED":"New Background applied",
     "BACKGROUND_DELETED":"Background deleted",
-	"ALBUMART_SETTINGS":"Album art Settings",
-	"WEB_ALBUMART_ENABLE_DOC":"Enable retrieval of Album art from Web, if no image is found in the file folder",
-	"WEB_ALBUMART_ENABLE":"Web Album art",
-	"WEB_ALBUMART_SIZE_DOC":"Choose the size of Album art downloaded from web: higher quality requires longer load times. Lower values may improve load times.",
-	"WEB_ALBUMART_SIZE":"Web Album art Size",
-	"SMALL":"Small",
-	"MEDIUM":"Medium",
-	"LARGE":"Large",
-	"EXTRALARGE":"Extra-large",
-	"MEGA":"Mega",
+    "ALBUMART_SETTINGS":"Album art Settings",
+    "WEB_ALBUMART_ENABLE_DOC":"Enable retrieval of Album art from Web, if no image is found in the file folder",
+    "WEB_ALBUMART_ENABLE":"Web Album art",
+    "WEB_ALBUMART_SIZE_DOC":"Choose the size of Album art downloaded from web: higher quality requires longer load times. Lower values may improve load times.",
+    "WEB_ALBUMART_SIZE":"Web Album art Size",
+    "SMALL":"Small",
+    "MEDIUM":"Medium",
+    "LARGE":"Large",
+    "EXTRALARGE":"Extra-large",
+    "MEGA":"Mega",
     "MUSIC_LIBRARY_SETTINGS":"Music Library Settings",
     "SHOW_TRACKNUMBERS":"Show Track numbers",
     "SHOW_TRACKNUMBERS_DOC":"Prepend track numbers to tracks title",
@@ -352,5 +352,9 @@
     "UPDATE":"Update",
     "CHECKING_FOR_UPDATES":"Checking for Updates",
     "CHECKING_FOR_UPDATES_WAIT":"Please Wait"
+  },
+  "TRANSLATION" : {
+    "TRANSLATION_MANAGER": "Translation Manager",
+    "LABEL": "Custom Translation"
   }
 }
