+++ app/plugins/miscellanea/translation_manager/config.json
@@ -1,0 +1,9 @@
+{
+  "language":{"value":"English","type":"string"},
+  "language_code":{"value":"en","type":"string"},
+  "background_type":{"value":"background","type":"string"},
+  "background_title":{"value":"Initial","type":"string"},
+  "background_path":{"value":"default.png","type":"string"},
+  "background_color":{"value":"#000000","type":"string"},
+  "theme":{"value":"default","type":"string"}
+}
