+++ app/plugins/miscellanea/translation_manager/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "translation_manager",
+  "version": "1.0.0",
+  "description": "Translation Manager plugin for Volumio2",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "volumio_info": {
+    "plugin_type": "miscellanea",
+    "boot_priority":10
+  }
+}
